requirements.txt
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
fredapi==0.5.2
frozendict==2.4.6
html5lib==1.1
idna==3.10
joblib==1.4.2
lxml==5.3.0
multitasking==0.0.11
numpy==2.2.0
pandas==2.2.3
pandas-datareader==0.10.0
peewee==3.17.8
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
soupsieve==2.6
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.2.50
